Send a PartyDescription object when registering a party
